Create VS 2010 based extension
This is a vanilla 2010 VSIX extension.  It was created with the project
wizard using the "Editor Margin" extension.  At this point the VSIX is
completely tied to Visual Studio 2010.  It can only be edited in and
deployed to that version.

Any attempt to open this project in 2012 or 2013 will be met with a
forced upgrade dialog.   Installing to either of these will also fail.

Over the next few edits this project will be updated such that it can be
edited in and deployed to any version of Visual Studio