Support VSIX deployment to any version of Visual Studio
The first change we will make is to enable the deployment of this
extension to any version of Visual Studio.  This is a relatively simple
process of updating the manifest to include the new versions of Visual
Studio in the SupprotedProducts section.

Once this change is added and the project rebuilt the resulting VSIX can
be installed on any version of Visual Studio

A quick note on Visual Studio versions

2010 = 10.0
2012 = 11.0
2013 = 12.0