Support project loading in any version of Visual Studio
This change will allow the RoundTripVSIX project to be loaded in any
version of Visual Studio.  This will be done without Visual Studio
attempting to upgrade the project.

The change really consists of two parts.  The first is the setting of
the MinimumVisualStudioVersion.  This needs to be done once per version
of Visual Studio beyond 2010.  The MinimumVisualStudioVersion property
doesn't affect the build in any meaningful way that I'm aware of.  But
the upgrade wizard does look for this when deciding whether or not an
upgrade is required.

The second change is to choose the SDK targets file based on the current
version of Visual Studio.  This is done by keying off the
$(VisualStudioVersion) property to build up the $(VsSdkTargets)
property.  The name of this property isn't important (other than
choosing one that Visual Studio isn't using).  The import of the SDK is
then switched to use this variable and be conditional on the SDK
existing.  The Conditional attribute is a must have here.  If it is
missing the upgrade wizard will flag this project for needing an upgrade

There are still several problems with the project though.  In particular
it still identifies the Visual Studio assemblies it references by a
non-strong name.  This means a build in 2012 will bind to the 2012
version of the assemblies.  These are built for the 4.5 framework and
hence can't be used in a 4.0 project.  This will result in a lot of
build errors.  It also has other problems building if the 2010 SDK isn't
installed.  The next commit will begin to take care of these problems