Switch to strongly named references
The RoundTripVSIX project can now be edited in a variety of Visual
Studio environments.  Each of these environments want to provide
different versions of the SDK references we are using.

It's important to guarantee that we are always using the reference
assemblies shipped with Visual Studio 2010.  These are fully supported
when loaded in newer versions (via redirects in devenv.exe.config).
Also all of the newer references are bound to newer versions of the .Net
Framework.  The RoundTripVSIX project will always target 4.0 hence even
if the new ones are bound to we can't reference them.

With this change we can now build the extension in any version of Visual
Studio so long as the 2010 SDK is installed on the machine.  That's
still pretty limiting though because the SDK requires a full Visual
Studio install to work.  The next commit will fix by deploying reference
assemblies with the source