Allow for build without full SDK installation
In this change we deploy the reference assemblies which are needed to
build the RoundTripVSIX project.   These are the reference assemblies
from a 2010 installation. All versions of Visual Studio beyond 2010 have
appropriate binding redirects for these values in the devenv.exe.config
file.  Hence our extension will load just fine if it references these.
This is generally true for any assembly that Visual Studio includes in
the SDK

The only change necessary to the project is to include this directory in
the reference search path.