Enable F5 debug everywhere
The final change we need is a small tweak to the .csproj so that F5
works out of the box.  This is done by adding the appropriate start
action to the .csproj to run devenv under a different registry hive.

Visual Studio generates this information in new VSIX projects but it
does so in the .csproj.user file.  Virtually every source code control
system excludes this file by default hence it will work until you clone
the source onto a new machine.  By putting this directly in the .csproj
file we ensure F5 will work for anyone who clones our source code